# Relyf.Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Relyf API",
    "description": "API for the Relyf upcycling platform - connecting users with creative ways to reuse and repurpose items",
    "version": "v1",
    "contact": {
      "name": "Relyf Team",
      "url": "https://github.com/nanasarf/Relyf"
    }
  },
  "servers": [
    {
      "url": "https://localhost:7139",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/Images": {
      "post": {
        "tags": ["Images"],
        "summary": "Upload image (JSON mode)",
        "description": "Upload image via base64 data URL, external URL, or Cloudinary",
        "operationId": "AddImage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddImageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "imageId": {
                      "type": "integer"
                    },
                    "url": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/Images/upload": {
      "post": {
        "tags": ["Images"],
        "summary": "Upload image (multipart form-data)",
        "description": "Upload image file via multipart/form-data (recommended method)",
        "operationId": "UploadImageFile",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file (max 10MB)"
                  },
                  "ownerType": {
                    "type": "string",
                    "enum": ["Item", "Idea", "Project"],
                    "description": "Type of owner entity"
                  },
                  "ownerId": {
                    "type": "integer",
                    "description": "ID of the owner entity"
                  },
                  "altText": {
                    "type": "string",
                    "description": "Alternative text for accessibility"
                  }
                },
                "required": ["file", "ownerType", "ownerId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "imageId": {
                      "type": "integer"
                    },
                    "url": {
                      "type": "string"
                    },
                    "fileName": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file or input"
          }
        }
      }
    },
    "/api/Images/{ownerType}/{ownerId}": {
      "get": {
        "tags": ["Images"],
        "summary": "Get images for an owner",
        "description": "List all images for a specific owner (Item, Idea, or Project)",
        "operationId": "ListImages",
        "parameters": [
          {
            "name": "ownerType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["Item", "Idea", "Project"]
            }
          },
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of images",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageRecord"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid owner type"
          }
        }
      }
    },
    "/api/Images/{imageId}": {
      "delete": {
        "tags": ["Images"],
        "summary": "Delete image",
        "description": "Remove an image from the system",
        "operationId": "DeleteImage",
        "parameters": [
          {
            "name": "imageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Image deleted successfully"
          },
          "404": {
            "description": "Image not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddImageRequest": {
        "type": "object",
        "required": ["ownerType", "ownerId"],
        "properties": {
          "ownerType": {
            "type": "string",
            "enum": ["Item", "Idea", "Project"],
            "description": "Type of entity that owns this image"
          },
          "ownerId": {
            "type": "integer",
            "description": "ID of the owner entity"
          },
          "source": {
            "type": "string",
            "nullable": true,
            "description": "Can be: 'upload', 'url', 'cloudinary', or a base64 data URL"
          },
          "url": {
            "type": "string",
            "nullable": true,
            "description": "External URL or will be generated for uploads"
          },
          "altText": {
            "type": "string",
            "nullable": true,
            "description": "Alternative text for accessibility"
          }
        }
      },
      "ImageRecord": {
        "type": "object",
        "properties": {
          "imageId": {
            "type": "integer"
          },
          "ownerType": {
            "type": "string"
          },
          "ownerId": {
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "altText": {
            "type": "string",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      }
    }
  },
...
}